Enhancement Update Product table

+++ src/main/java/com/gabr/ecommerce/controllers/CartController.java
@@ -14,24 +14,24 @@ public class CartController {
     private final CartService cartService;
 
     @GetMapping("/{userId}")
-    public ResponseEntity<ApiResponse<CartDto>> getCart(@PathVariable int userId) {
+    public ResponseEntity<ApiResponse<CartDto>> getCart(@PathVariable Long userId) {
         return ResponseEntity.ok(ApiResponse.success("Cart fetched", cartService.getUserCart(userId)));
     }
 
     @PostMapping("/{userId}/add/{productId}")
     public ResponseEntity<ApiResponse<CartDto>> addItem(
-            @PathVariable int userId,
+            @PathVariable Long userId,
             @PathVariable Long productId,
             @RequestParam(defaultValue = "1") int quantity) {
         return ResponseEntity.ok(ApiResponse.success("Item added", cartService.addItem(userId, productId, quantity)));
     }
 
     @DeleteMapping("/{userId}/remove/{productId}")
-    public ResponseEntity<ApiResponse<CartDto>> removeItem(@PathVariable int userId, @PathVariable Long productId) {
+    public ResponseEntity<ApiResponse<CartDto>> removeItem(@PathVariable Long userId, @PathVariable Long productId) {
         return ResponseEntity.ok(ApiResponse.success("Item removed", cartService.removeItem(userId, productId)));
     }
     @DeleteMapping("/{userId}/clear")
-    public ResponseEntity<ApiResponse<CartDto>> clearCart(@PathVariable int userId) {
+    public ResponseEntity<ApiResponse<CartDto>> clearCart(@PathVariable Long userId) {
         return ResponseEntity.ok(ApiResponse.success("Cart cleared", cartService.clearCart(userId)));
     }
