+++ src/main/java/com/gabr/ecommerce/service/CartService.java
@@ -3,8 +3,8 @@ package com.gabr.ecommerce.service;
 import com.gabr.ecommerce.dto.CartDto;
 
 public interface CartService {
-    CartDto getUserCart(int userId);
-    CartDto addItem(int userId, Long productId, int quantity);
-    CartDto removeItem(int userId, Long productId);
-    CartDto clearCart(int userId);
+    CartDto getUserCart(Long userId);
+    CartDto addItem(Long userId, Long productId, int quantity);
+    CartDto removeItem(Long userId, Long productId);
+    CartDto clearCart(Long userId);
 }
