+++ src/main/java/com/gabr/ecommerce/service/OrderService.java
@@ -4,7 +4,7 @@ import com.gabr.ecommerce.dto.OrderDto;
 import java.util.List;
 
 public interface OrderService {
-    OrderDto placeOrder(int userId);
-    List<OrderDto> getUserOrders(int userId);
+    OrderDto placeOrder(Long userId);
+    List<OrderDto> getUserOrders(Long userId);
     OrderDto getOrderById(Long orderId);
 }
